Convert the main class to Plugin and extract the ManagementLink
