+++ src/main/resources/com/cloudbees/simplediskusage/QuickDiskUsagePlugin/index.jelly
